requirements.txt
python-dotenv==0.20.0
fastapi==0.75.1
uvicorn==0.15.0
responses==0.20.0
pytest==7.1.1